requirements.txt
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pymysql==1.1.0
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
ta==0.11.0
beautifulsoup4==4.12.2
requests==2.31.0
textblob==0.17.1
nltk==3.8.1